LayoutTests/platform/ios-simulator/accessibility/non-contiguous-link-expected.txt

ALERT: Tests passed
This is a link that would normally appear as multiple elements.
